Add a lowpass filter helper
The lowpass filter is given the latest value to add and returns the current
lowpass filtered version. Also needed is an 'order' parameter which controls
the strength of the filter (higher is slower response, stronger filtering),
and a scratch area to keep track of internal state in between invocations.

This is the simplest of the 'leaky integrator' lowpass filters.