Change the data peak finder to ignore zero passings and idle states.
Peak finder was finding peaks in the noise, and also getting confused when the
z accel is large negative at the beginning, as it will never cross the zero
axis. Keep track of the max and min per peak (detected via transitions on
the derivative's sign), and use midpoint crossing instead.